requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
openpyxl==3.1.2
pytz==2024.1
python-dotenv==1.0.0
urllib3==2.1.0
